Coordinates struct and refactoring
